+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.4.0-dev.53](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.4.0-dev.52...v1.4.0-dev.53) (2026-04-02)
+
+
+### Bug Fixes
+
+* Manage Page Table Header should stick when scroll down ([073d211](https://github.com/michael-raquel/PowerIntake-WebApp/commit/073d2114911ae167f8842e5b2650d5a0d7b37380))
+
+# [1.4.0-dev.52](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.4.0-dev.51...v1.4.0-dev.52) (2026-04-02)
+
+
+### Bug Fixes
+
+* sticky table + search filter in navbar ([e353b51](https://github.com/michael-raquel/PowerIntake-WebApp/commit/e353b51abf26737e0d9c8edc0e69a3faafef2e14))
+
 # [1.4.0-dev.51](https://github.com/michael-raquel/PowerIntake-WebApp/compare/v1.4.0-dev.50...v1.4.0-dev.51) (2026-04-02)